+++ esp-hal/src/soc/esp32s2/mod.rs
@@ -24,21 +24,6 @@ pub mod gpio;
 pub mod peripherals;
 pub(crate) mod regi2c;
 
-/// The name of the chip ("esp32s2") as `&str`
-#[macro_export]
-macro_rules! chip {
-    () => {
-        "esp32s2"
-    };
-}
-
-/// A link to the Technical Reference Manual (TRM) for the chip.
-#[doc(hidden)]
-#[macro_export]
-macro_rules! trm_link {
-    () => { "https://www.espressif.com/sites/default/files/documentation/esp32-s2_technical_reference_manual_en.pdf" };
-}
-
 #[cfg_attr(not(feature = "unstable"), allow(unused))]
 pub(crate) mod constants {
     use crate::time::Rate;
@@ -47,10 +32,6 @@ pub(crate) mod constants {
     pub const I2S_SCLK: u32 = 160_000_000;
     /// Default clock source for the I2S peripheral.
     pub const I2S_DEFAULT_CLK_SRC: u32 = 2;
-    /// Start address of the RMT (Remote Control) peripheral's RAM.
-    pub const RMT_RAM_START: usize = 0x3f416400;
-    /// The size (number of pulse codes) of each RMT channel's dedicated RAM.
-    pub const RMT_CHANNEL_RAM_SIZE: usize = 64;
     /// Reference clock tick frequency, set to 1 MHz.
     pub const REF_TICK: Rate = Rate::from_mhz(1);
 }