+++ esp-metadata/src/generate_cfg.rs
@@ -2,6 +2,7 @@ use core::str::FromStr;
 use std::sync::OnceLock;
 
 use anyhow::{Result, bail};
+use proc_macro2::TokenStream;
 use strum::IntoEnumIterator;
 
 macro_rules! include_toml {
@@ -185,12 +186,37 @@ pub struct MemoryRegion {
 
 #[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
 struct Device {
-    pub name: String,
-    pub arch: Arch,
-    pub cores: Cores,
-    pub peripherals: Vec<String>,
-    pub symbols: Vec<String>,
-    pub memory: Vec<MemoryRegion>,
+    name: String,
+    arch: Arch,
+    cores: usize,
+    trm: String,
+    peripherals: Vec<String>,
+    symbols: Vec<String>,
+    memory: Vec<MemoryRegion>,
+    #[serde(default)]
+    rmt: Option<RmtProperties>,
+}
+
+#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
+struct RmtProperties {
+    ram_start: u32,
+    channel_ram_size: u32,
+}
+
+impl RmtProperties {
+    fn properties(&self) -> impl Iterator<Item = (&str, TokenStream)> {
+        [
+            ("ram_start", number(self.ram_start)),
+            ("channel_ram_size", number(self.channel_ram_size)),
+        ]
+        .into_iter()
+    }
+}
+
+// Output a Display-able value as a TokenStream, intended to generate numbers
+// without the type suffix.
+fn number(n: impl std::fmt::Display) -> TokenStream {
+    TokenStream::from_str(&format!("{n}")).unwrap()
 }
 
 /// Device configuration file format.
@@ -219,10 +245,12 @@ impl Config {
             device: Device {
                 name: "".to_owned(),
                 arch: Arch::RiscV,
-                cores: Cores::Single,
+                cores: 1,
+                trm: "".to_owned(),
                 peripherals: Vec::new(),
                 symbols: Vec::new(),
                 memory: Vec::new(),
+                rmt: None,
             },
         }
     }
@@ -239,7 +267,11 @@ impl Config {
 
     /// The core count of the device.
     pub fn cores(&self) -> Cores {
-        self.device.cores
+        if self.device.cores > 1 {
+            Cores::Multi
+        } else {
+            Cores::Single
+        }
     }
 
     /// The peripherals of the device.
@@ -265,7 +297,10 @@ impl Config {
         [
             self.device.name.as_str(),
             self.device.arch.as_ref(),
-            self.device.cores.as_ref(),
+            match self.cores() {
+                Cores::Single => "single_core",
+                Cores::Multi => "multi_core",
+            },
         ]
         .into_iter()
         .chain(self.device.peripherals.iter().map(|s| s.as_str()))
@@ -288,18 +323,81 @@ impl Config {
         // Define env-vars for all memory regions
         for memory in self.memory() {
             println!("cargo:rustc-cfg=has_{}_region", memory.name.to_lowercase());
-
-            println!(
-                "cargo::rustc-env=ESP_METADATA_REGION_{}_START={}",
-                memory.name.to_uppercase(),
-                memory.start
-            );
-            println!(
-                "cargo::rustc-env=ESP_METADATA_REGION_{}_END={}",
-                memory.name.to_uppercase(),
-                memory.end
-            );
         }
+    }
+
+    pub fn generate_metadata(&self) {
+        let out_dir = std::env::var_os("OUT_DIR").unwrap();
+        let out_dir = std::path::Path::new(&out_dir);
+        let out_file = out_dir.join("_generated.rs").to_string_lossy().to_string();
+
+        let mut g = TokenStream::new();
+
+        let chip_name = self.name();
+        // Public API, can't use a private macro:
+        g.extend(quote::quote! {
+            /// The name of the chip as `&str`
+            #[macro_export]
+            macro_rules! chip {
+                () => { #chip_name };
+            }
+        });
+
+        // Translate the chip properties into a macro that can be used in esp-hal:
+        let arch = self.device.arch.as_ref();
+        let cores = number(self.device.cores);
+        let trm = &self.device.trm;
+
+        let peripheral_properties = std::iter::once(
+            self.device.rmt.iter().map(|rmt| ("rmt", rmt.properties())),
+        )
+        .flat_map(|groups| {
+            groups.flat_map(move |(group, group_properties)| {
+                group_properties.map(move |(name, value)| {
+                    let name = format!("{group}.{name}");
+                    quote::quote! {
+                        (#name) => { #value };
+                    }
+                })
+            })
+        });
+
+        // Not public API, can use a private macro:
+        g.extend(quote::quote! {
+            /// A link to the Technical Reference Manual (TRM) for the chip.
+            #[doc(hidden)]
+            #[macro_export]
+            macro_rules! property {
+                ("chip") => { #chip_name };
+                ("arch") => { #arch };
+                ("cores") => { #cores };
+                ("cores", str) => { stringify!(#cores) };
+                ("trm") => { #trm };
+                #(#peripheral_properties)*
+            }
+        });
+
+        let region_branches = self.memory().iter().map(|region| {
+            let name = region.name.to_uppercase();
+            let start = number(region.start as usize);
+            let end = number(region.end as usize);
+
+            quote::quote! {
+                ( #name ) => {
+                    #start .. #end
+                };
+            }
+        });
+
+        g.extend(quote::quote! {
+            /// Macro to get the address range of the given memory region.
+            #[macro_export]
+            macro_rules! memory_range {
+                #(#region_branches)*
+            }
+        });
+
+        std::fs::write(&out_file, g.to_string()).unwrap();
     }
 }